extracted tailwind v1 colors into separate file
